Minor changes
	* Quartz added, for time scheduled tasks
	* Exception and Attributes projects are moved under Infrastructure projects

+++ MS.Katusha.Domain.Raven/Entities/ConversationResult.cs
@@ -2,8 +2,8 @@ namespace MS.Katusha.Domain.Raven.Entities
 {
     public class ConversationResult
     {
-        public int FromId { get; set; }
-        public int ToId { get; set; }
+        public long FromId { get; set; }
+        public long ToId { get; set; }
         public int Count { get; set; }
         public int UnreadCount { get; set; }
     }
